agent_companion_cli/pyproject.toml
[project]
name = "agent-companion-cli"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = ["watchdog>=4.0.1","psutil>=5.9.8","pyyaml>=6.0.2"]

[project.scripts]
agent-companion = "agent_companion.main:main"
